fix delete cover, fix delete song cascade song playlist

+++ src/Application/Command/Song/DeleteSong/DeleteSongHandler.php
@@ -3,16 +3,33 @@
 namespace App\Application\Command\Song\DeleteSong;
 
 use App\Domain\Repository\DomainSongRepositoryInterface;
+use Symfony\Component\HttpKernel\KernelInterface;
 
 class DeleteSongHandler
 {
+    private KernelInterface $kernel;
+    private DomainSongRepositoryInterface $songRepository;
+
     public function __construct(
-        public DomainSongRepositoryInterface $songRepository,
+        KernelInterface $kernel,
+        DomainSongRepositoryInterface $songRepository
     ) {
+        $this->kernel = $kernel;
+        $this->songRepository = $songRepository;
     }
+
     public function __invoke(DeleteSong $deleteSong): void
     {
         $song = $this->songRepository->findOneBy(['id' => $deleteSong->songId]);
-        $this->songRepository->remove($song, true);
+
+        if ($song) {
+            $coverPath = $this->kernel->getProjectDir() . '/public/songs/covers/' . $song->getPhotoAlbum();
+
+            if (file_exists($coverPath)) {
+                unlink($coverPath);
+            }
+
+            $this->songRepository->remove($song, true);
+        }
     }
 }
